+++ samples/SqlServerDemo/SqlServerDemo.Database/Migrations/20211207-213319-02-create-ntangle-versiontracking-table.sql
@@ -1,9 +1,8 @@
 CREATE TABLE [NTangle].[VersionTracking] (
   [VersionTrackingId] UNIQUEIDENTIFIER NOT NULL DEFAULT (NEWSEQUENTIALID()) PRIMARY KEY NONCLUSTERED ([VersionTrackingId] ASC),
-  [Schema] VARCHAR(64) NOT NULL,
-  [Table] VARCHAR(128) NOT NULL,
-  [Key] NVARCHAR(128) NOT NULL,
+  [Object] NVARCHAR(192) NOT NULL,
+  [Key] NVARCHAR(256) NOT NULL,
   [Hash] NVARCHAR(128) NOT NULL,
   [BatchTrackingId] BIGINT NOT NULL,
-  CONSTRAINT [IX_NTangle_VersionTracking_SchemaTableKey] UNIQUE CLUSTERED ([Schema], [Table], [Key])
+  CONSTRAINT [IX_NTangle_VersionTracking_SchemaTableKey] UNIQUE CLUSTERED ([Object], [Key])
 );